third_party/WebKit/LayoutTests/platform/android/fast/events/window-onerror-12-expected.txt

CONSOLE MESSAGE: line 17: 1
CONSOLE MESSAGE: line 17: 2
CONSOLE MESSAGE: line 17: 3
This test should trigger 'window.onerror' for the exception in the attribute handler, regardless of how it's set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.onerror: "Uncaught TypeError: Cannot call method 'mu' of null" at window-onerror-12.html (Line: 12, Column: 75)
Stack Trace:
TypeError: Cannot call method 'mu' of null
    at HTMLButtonElement.onclick window-onerror-12.html:12:123
    at window-onerror-12.html:33:24

Returning 'true': the error should not be reported in the console as an unhandled exception.



window.onerror: "Uncaught TypeError: Cannot call method 'mu' of null" at window-onerror-12.html (Line: 35, Column: 75)
Stack Trace:
TypeError: Cannot call method 'mu' of null
    at HTMLButtonElement.onclick window-onerror-12.html:35:104
    at window-onerror-12.html:33:24

Returning 'true': the error should not be reported in the console as an unhandled exception.



window.onerror: "Uncaught TypeError: Cannot call method 'mu' of null" at window-onerror-12.html (Line: 35, Column: 75)
Stack Trace:
TypeError: Cannot call method 'mu' of null
    at HTMLButtonElement.onclick window-onerror-12.html:35:104
    at window-onerror-12.html:33:24

Returning 'true': the error should not be reported in the console as an unhandled exception.



PASS successfullyParsed is true

TEST COMPLETE
Button.  Button 2 Button 3
